move: study 관련 dto request, response 패키지 생성 후 이동
