+++ src/main/java/com/web/stard/domain/board/global/application/StarScrapService.java
@@ -4,7 +4,7 @@ import com.web.stard.domain.board.global.domain.enums.ActType;
 import com.web.stard.domain.board.global.domain.enums.TableType;
 import com.web.stard.domain.board.global.dto.response.PostResponseDto;
 import com.web.stard.domain.member.domain.Member;
-import com.web.stard.domain.study.domain.dto.StudyResponseDto;
+import com.web.stard.domain.study.domain.dto.response.StudyResponseDto;
 
 public interface StarScrapService {
     Long addStarScrap(Member member, Long targetId, ActType actType, TableType tableType);
